Organizando backend y modoficando endpoins para hacerlos más descriptivos
