+++ backend/src/routes/controller/courses/getAllDataCourses.js
@@ -1,0 +1,27 @@
+//Función para obtener la data de todos los cursos
+const { Category, Course, Review } = require("../../../db.js");
+const { getInfoCourse } = require("./getInfoCourse");
+
+const getAllDataCourses = async () => {
+  try {
+    let getAllCourses = await Course.findAll({
+      //Busca todos los cursos
+      attributes: ["name"],
+    });
+    let arrayAllCoursesInfo = []; //Array que contendrá todos los cursos
+    for (const courseName of getAllCourses) {
+      //Recorre todos los cursos
+      let temporaryInfo = await getInfoCourse(courseName.dataValues.name); //Obtiene la información del curso
+      arrayAllCoursesInfo.push(temporaryInfo); //Agrega la información del curso al array
+    }
+    return arrayAllCoursesInfo;
+    // getInfoCourse(name)
+  } catch (error) {
+    console.error(error);
+    return { message: "Error al obtener los cursos" };
+  }
+};
+
+module.exports = {
+  getAllDataCourses,
+};
