+++ backend/src/routes/controller/courses/update.course.controller.js
@@ -1,5 +1,5 @@
 const { Course, Category } = require("../../../db");
-const { getInfoCourse } = require("./courses.controller.js");
+const { getInfoCourse } = require("./get.courses.controller.js");
 const { getCategoryId } = require("../getCategoryId");
 const updateCourse = async (req, res) => {
   const { id } = req.params;
