+++ backend/src/routes/courses.js
@@ -1,18 +1,14 @@
 const { Router } = require("express");
 const router = Router();
-const {
-  getCourses,
-  getInfoCourse,
-  getCourseDetail,
-} = require("./controller/courses/courses.controller.js");
-const { postCourses } = require("./controller/courses/post.courses.js");
-const { deleteCourse } = require("./controller/courses/delete.course.js");
-const { updateCourse } = require("./controller/courses/update.course.js");
+const { getCourses, getCourseDetail } = require("./controller/courses/get.courses.controller.js");
+const { postCourses } = require("./controller/courses/post.courses.controller.js");
+const { deleteCourse } = require("./controller/courses/delete.course.controller.js");
+const { updateCourse } = require("./controller/courses/update.course.controller.js");
 
-router.post("/", postCourses); //create a course
+router.post("/create", postCourses); //create a course
 router.get("/", getCourses); //trae todos los cursos,si tiene query filtra esos cursos
-router.get("/:id", getCourseDetail); //trae todos los cursos,si tiene query filtra esos cursos
-router.delete("/:id", deleteCourse);
-router.put("/:id", updateCourse);
+router.get("/detail/:id", getCourseDetail); //trae todos los cursos,si tiene query filtra esos cursos
+router.delete("/delete/:id", deleteCourse);
+router.put("/update/:id", updateCourse);
 
 module.exports = router;
