+++ backend/src/routes/students.js
@@ -5,13 +5,13 @@ const router = Router();
 const {
   getStudents,
   getStudent,
-  updateStudent,
-  deteteStudent,
-} = require("./controller/students/getStudents");
+} = require("./controller/students/get.students.controller");
+const { updateStudent } = require("./controller/students/update.student.controller")
+const { deteteStudent } = require("./controller/students/delete.student.controller")
 
 router.get("/", getStudents);
-router.get("/:id", getStudent);
-router.put("/:id", updateStudent);
-router.delete("/:id", deteteStudent);
+router.get("/detail/:id", getStudent);
+router.put("/update/:id", updateStudent);
+router.delete("/delete/:id", deteteStudent);
 
 module.exports = router;
