+++ backend/src/routes/teachers.js
@@ -4,13 +4,13 @@ const router = Router();
 const {
   getTeachers,
   getTeacher,
-  updateTeacher,
-  deteteTeacher,
-} = require("./controller/teachers/getTeachers");
+} = require("./controller/teachers/get.teacher.controller");
+const { updateTeacher } = require("./controller/teachers/update.teacher.controller")
+const { deteteTeacher } = require("./controller/teachers/delete.teacher.controller")
 
 router.get("/", getTeachers);
-router.get("/:id", getTeacher);
-router.put("/:id", updateTeacher);
-router.delete("/:id", deteteTeacher);
+router.get("/detail/:id", getTeacher);
+router.put("/update/:id", updateTeacher);
+router.delete("/delete/:id", deteteTeacher);
 
 module.exports = router;
